// backend-api/src/third-party-apis/models/get_crawled_id.response.ts
export interface GetCrawledIdResponse {
  version: string;
  status_code: number;
  status_message: string;
  time: string;
  cost: number;
  tasks_count: number;
  tasks_error: number;
  tasks: Task[];
}

export interface Task {
  id: string;
  status_code: number;
  status_message: string;
  time: string;
  cost: number;
  result_count: number;
  path: string[];
  data: Data;
  result: null;
}

export interface Data {
  api: string;
  function: string;
  target: string;
  max_crawl_pages: number;
}