// backend-api/src/third-party-apis/models/website_speed.response.ts
export interface WebsiteSpeedResponse {
  version: string;
  status_code: number;
  status_message: string;
  time: string;
  cost: number;
  tasks_count: number;
  tasks_error: number;
  tasks: DemoTask[];
}

export interface DemoTask {
  id: string;
  status_code: number;
  status_message: string;
  time: string;
  cost: number;
  result_count: number;
  path: string[];
  data: TaskData;
  result: Result[];
}

export interface TaskData {
  api: string;
  function: string;
  url: string;
  for_mobile: boolean;
  categories: string[];
  audits: string[];
}

export interface Result {
  lighthouseVersion: string;
  requestedUrl: string;
  mainDocumentUrl: string;
  finalDisplayedUrl: string;
  finalUrl: string;
  fetchTime: Date;
  gatherMode: string;
  runWarnings: any[];
  userAgent: string;
  environment: Environment;
  audits: Audits;
  configSettings: ConfigSettings;
  categories: Categories;
  categoryGroups: CategoryGroups;
  stackPacks: StackPack[];
  entities: Entity[];
  fullPageScreenshot: FullPageScreenshot;
  timing: Timing;
  i18n: I18N;
}